+++ CHANGES.md
@@ -13,8 +13,12 @@
 
 ## 2022.3.0
 
+- [CHANGE] `--hatch-rate` を `--vcs-hatch-rate` に変更
+    - @melpon
 - [UPDATE] WebRTC を m102.5005.7.6 に上げる
     - @melpon @voluntas
+- [UPDATE] WebRTC を m102.5005.7.1 に上げる
+    - @melpon
 - [UPDATE] asmjit と blend2d を最新版に上げる
     - @voluntas @melpon
 - [UPDATE] `cmake` を `3.23.2` に上げる
@@ -30,6 +34,8 @@
 - [ADD] spotlight_focus_rid を追加する
     - @melpon
 - [ADD] spotlight_unfocus_rid を追加する
+    - @melpon
+- [ADD] `--instance-hatch-rate` を追加
     - @melpon
 
 ## 2022.2.0